Updating to BigNumber for most numbers

+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/api/models/ApiAccount.kt
@@ -3,7 +3,7 @@ package com.dllewellyn.coinbaseapi.api.models
 import com.dllewellyn.coinbaseapi.models.Account
 import com.dllewellyn.coinbaseapi.models.currency.SupportedCurrency
 import com.google.gson.annotations.SerializedName
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class ApiAccount(
     @SerializedName("available") val available: String,
@@ -15,9 +15,9 @@ data class ApiAccount(
 ) {
     fun toCore() = Account(
         SupportedCurrency(currency, currency, 0.0),
-        BigInteger(balance),
-        BigInteger(available),
-        BigInteger(hold),
+        BigDecimal(balance),
+        BigDecimal(available),
+        BigDecimal(hold),
         id
     )
 }
