+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/ExchangeRates.kt
@@ -1,8 +1,8 @@
 package com.dllewellyn.coinbaseapi.models
 
 import com.dllewellyn.coinbaseapi.models.currency.CryptoCurrency
-import java.math.BigInteger
+import java.math.BigDecimal
 
-data class ExchangeRates(val cryptoCurrency: CryptoCurrency, val map : Map<String, BigInteger>) {
+data class ExchangeRates(val cryptoCurrency: CryptoCurrency, val map : Map<String, BigDecimal>) {
     fun filterForCurrency(currency : String) = map[currency]
 }
