+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/marketinfo/ProductTicker.kt
@@ -1,10 +1,10 @@
 package com.dllewellyn.coinbaseapi.models.marketinfo
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class ProductTicker(
-    val ask: BigInteger,
-    val bid: BigInteger,
+    val ask: BigDecimal,
+    val bid: BigDecimal,
     val price: String,
     val size: String,
     val time: String,
