+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/trade/OrderRequest.kt
@@ -2,7 +2,7 @@ package com.dllewellyn.coinbaseapi.models.trade
 
 import com.dllewellyn.coinbaseapi.models.BuyOrSell
 import com.dllewellyn.coinbaseapi.models.currency.CurrencyPair
-import java.math.BigInteger
+import java.math.BigDecimal
 
-class OrderRequest(val buyOrSell: BuyOrSell, val currency: CurrencyPair, val size: BigInteger, val price: BigInteger,
+class OrderRequest(val buyOrSell: BuyOrSell, val currency: CurrencyPair, val size: BigDecimal, val price: BigDecimal,
                    val type : LimitOrMarket)
