+++ coroutines-adapter/src/main/kotlin/com/dllewellyn/coinbaseapi/adapter/CurrencyPriceAdapter.kt
@@ -8,7 +8,7 @@ import com.dllewellyn.coinbaseapi.models.currency.OrderBookLevel
 import com.dllewellyn.coinbaseapi.models.currency.OrderBookList
 import com.dllewellyn.coinbaseapi.models.currency.OrderFromBook
 import com.dllewellyn.coinbaseapi.models.trade.CurrencyBuyAndSell
-import java.math.BigInteger
+import java.math.BigDecimal
 
 
 class CurrencyPriceAdapter(private val retrofit: RetrofitCoroutinesBuilder) : CurrencyPrice {
@@ -21,8 +21,8 @@ class CurrencyPriceAdapter(private val retrofit: RetrofitCoroutinesBuilder) : Cu
                 CurrencyBuyAndSell(
                     pair.baseCurrency,
                     pair.quoteCurrency,
-                    BigInteger(it.ask ?: "0"),
-                    BigInteger(it.bid ?: "0")
+                    BigDecimal(it.ask ?: "0"),
+                    BigDecimal(it.bid ?: "0")
                 )
             }
     }
@@ -35,8 +35,8 @@ class CurrencyPriceAdapter(private val retrofit: RetrofitCoroutinesBuilder) : Cu
                 val asks = orderBook.asks.map { asks ->
                     (asks as List<*>).let { ask ->
                         OrderFromBook(
-                            BigInteger(ask[0] as String),
-                            BigInteger(ask[1] as String),
+                            BigDecimal(ask[0] as String),
+                            BigDecimal(ask[1] as String),
                             (ask[2] as Double).toInt()
                         )
                     }
@@ -45,8 +45,8 @@ class CurrencyPriceAdapter(private val retrofit: RetrofitCoroutinesBuilder) : Cu
                 val bids = orderBook.bids.map { bids ->
                     (bids as List<*>).let { ask ->
                         OrderFromBook(
-                            BigInteger(ask[0] as String),
-                            BigInteger(ask[1] as String),
+                            BigDecimal(ask[0] as String),
+                            BigDecimal(ask[1] as String),
                             (ask[2] as Double).toInt()
                         )
                     }
